debate/chat/requirements.txt
google-genai
google-api-core
